Parse number from maxAge from options
If the maxAge from options are passed from an env variable it is very
likely it is a string. We now try to parse it and use the value as a
number.

The error message was unreadable, as it would cause an infinite recursive
loop ending in a maximum call stack size exceeded and ending the
process.